+++ src/lib/Form.tsx
@@ -23,7 +23,9 @@ interface FormProps<T extends FieldValues> {
   defaultValues?: DeepPartial<T>;
 
   /**
-   * passed fieldnames will be marked with "*"
+   * passed field names will be marked with "*"
+   * field paths for arrays are supported as 0-based indexes (e.g. "items.0", "items.0.name") in compliance with react-hook-form types
+   * @example ['firstName', 'address.street', 'items.0.name', 'items.0']
    */
   requiredFields?: FieldPath<T>[];