+++ CHANGELOG.md
@@ -7,13 +7,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- support into `requiredFields` property of `Form` component, for nested objects and arrays.
+- `form` helper functions.
+- `countryMenuWidth` property to `TelephoneNumberInput` in order to customize the country menu width.
+
 ### Fixed
 
 - `TelephoneNumberInput` countries order, in order to be alphabetically sorted.
+- Required field label on `FormGroupLayoutLabel`, `ColorPicker`, `TelephoneNumberInput`, `TypeaheadTextField` (hence `StaticTypeaheadInput` and `AsyncTypeaheadInput`) in order to display \* also on nested and array fields.
 
-### Added
+  1. `requiredFields` can still accept a `FieldPath<T>[]`
+  2. In order to be complaint with `FieldPath` react-hook-form type (`object.${number}.property`) array properties provide a wildcard:
 
-- `countryMenuWidth` property to `TelephoneNumberInput` in order to customize the country menu width.
+  ```tsx
+  requiredFields = [
+    `object`,
+    `object.nestedObjects`,
+    `objects.*.property`,
+    `object.nestedObject.property`,
+    `object.nestedObjects.*.property`,
+  ];
+  ```
+
+  is going to consider as required:
+
+  ```tsx
+    name="object"
+    name="object.nestedObjects.0", "object.nestedObjects.1", etc.
+    name="objects.0.property", name="objects.1.property", etc.
+    name="object.nestedObject.property"
+    name="object.nestedObjects.0.property", name="object.nestedObjects.1.property", etc.
+  ```
 
 ## [3.13.1] - 2025-11-13